pyproject.toml
[project]
name = "app-virgen-del-carmen"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlalchemy>=2",
  "pydantic>=2",
  "python-dotenv",
  "pandas",
  "openpyxl",
  "qrcode[pil]"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
